split redux actions into server and client
