+++ src/redux/actions/client/replaceState.ts
@@ -1,5 +1,5 @@
 import { createAction } from '@reduxjs/toolkit';
 
-import { type AppState } from '../store';
+import { type AppState } from '../../store';
 
 export const replaceState = createAction<AppState>('REPLACE_STATE');
