+++ src/redux/actions/index.ts
@@ -1,3 +1,2 @@
-export * from './replaceState';
-export * from './user';
-export * from './playlist';
+export * from './client/replaceState';
+export * from './server/user';
