+++ src/redux/actions/server/init.ts
@@ -2,8 +2,8 @@
 
 import { tryGetAuthSession } from '@/spotify/util';
 
-import store from '../store';
-import { setAuthStatus } from '../slices/userSlice';
+import store from '../../store';
+import { setAuthStatus } from '../../slices/userSlice';
 
 export const readAuthSession = (): void => {
   const authSession = tryGetAuthSession();
